seden: Release v3.2 - Bot tabanı iyileştirilmiştir. - Bilinen hataların büyük bir çoğunluğu çözülmüştür. - Hem SQL hem de NON-SQL olacak şekilde uygun uyarlamalar yapılmıştır. - Çakışan komutlar çözülmüştür.

+++ sedenbot/moduller/purge.py
@@ -21,7 +21,7 @@ from asyncio import sleep
 from telethon.errors import rpcbaseerrors
 
 from sedenbot import BOTLOG, BOTLOG_CHATID, CMD_HELP
-from sedenbot.events import sedenify
+from sedenbot.events import extract_args, sedenify
 
 @sedenify(outgoing=True, pattern="^.purge$")
 async def fastpurger(purg):
@@ -31,7 +31,7 @@ async def fastpurger(purg):
     itermsg = purg.client.iter_messages(chat, min_id=purg.reply_to_msg_id)
     count = 0
 
-    if purg.reply_to_msg_id is not None:
+    if purg.reply_to_msg_id :
         async for msg in itermsg:
             msgs.append(msg)
             count = count + 1
@@ -59,13 +59,15 @@ async def fastpurger(purg):
 @sedenify(outgoing=True, pattern="^.purgeme")
 async def purgeme(delme):
     """ .purgeme komutu belirtilen miktarda kullanıcın mesajlarını siler. """
-    message = delme.text
-    count = int(message[9:])
+    count = extract_args(delme)
+    if not count.isdigit():
+        await delme.edit("`Temizlik yapılamadı, sayı geçersiz.`")
+        return
     i = 1
 
     async for message in delme.client.iter_messages(delme.chat_id,
                                                     from_user='me'):
-        if i > count + 1:
+        if i > int(count) + 1:
             break
         i = i + 1
         await message.delete()
@@ -93,7 +95,7 @@ async def delete_it(delme):
             if BOTLOG:
                 await delme.client.send_message(
                     BOTLOG_CHATID, "Hedeflenen mesajın silinmesi başarılıyla tamamlandı")
-        except rpcbaseerrors.BadRequestError:
+        except:
             if BOTLOG:
                 await delme.client.send_message(
                     BOTLOG_CHATID, "Bu mesajı silemiyorum.")
@@ -104,7 +106,7 @@ async def editer(edit):
     message = edit.text
     chat = await edit.get_input_chat()
     self_id = await edit.client.get_peer_id('me')
-    string = str(message[6:])
+    string = extract_args(edit)
     i = 1
     async for message in edit.client.iter_messages(chat, self_id):
         if i == 2:
@@ -119,9 +121,13 @@ async def editer(edit):
 @sedenify(outgoing=True, pattern="^.sd")
 async def selfdestruct(destroy):
     """ .sd komutu kendi kendine yok edilebilir mesajlar yapar. """
-    message = destroy.text
-    counter = int(message[4:6])
-    text = str(destroy.text[6:])
+    args = extract_args(destroy)
+    argv = args.split(' ', 1)
+    if len(args) < 1 or len(argv) < 2 or not argv[0].isdigit():
+        await destroy.edit("`Komut kullanımı hatalı.`")
+        return
+    counter = int(argv[0])
+    text = argv[1]
     await destroy.delete()
     smsg = await destroy.client.send_message(destroy.chat_id, text)
     await sleep(counter)
